+++ src/hkimjp/wil2/my.clj
@@ -1,0 +1,8 @@
+(ns hkimjp.wil2.my
+  (:require
+   [ring.util.response :as response]
+   [hkimjp.wil2.view :refer [page]]))
+
+(defn my [request]
+  (page [:div "my"]))
+