+++ src/hkimjp/wil2/weeks.clj
@@ -1,0 +1,10 @@
+(ns hkimjp.wil2.weeks
+  (:require
+   [ring.util.response :as response]
+   [hkimjp.wil2.view :refer [page]]))
+
+(defn list-days [request]
+  (page [:div "list-days"]))
+
+(defn browse [request]
+  (page [:div "browse"]))